moved version checking code into a service
