+++ SendItems/SendItems.cs
@@ -10,7 +10,6 @@ using Denifia.Stardew.SendItems.Events;
 
 namespace Denifia.Stardew.SendItems
 {
-    // TODO: Merge this with ModEntry
     public class SendItems
     {
         private readonly IMod _mod;
